CAMEL-21641: Java DSL: RouteBuilder to support inlined component configurers using lambda style for low-code usage in Java to be in a single soure file, that supports jbang exporting (stub)
